commit 2
small steps, need to figure out how to run this in the gpu rather than the cpu

+++ SlimemoldSimulation/SlimemoldSimulation/slime.h
@@ -6,10 +6,10 @@
 class slime
 {
 private:
-	sf::Vector2i pos{ 0, 0 };
+	sf::Vector2f pos{ 0, 0 };
 	float angle{ 0 };
 public:
-	void initialise(sf::Vector2i p, float a);
+	void initialise(sf::Vector2f p, float a);
 	void update(sf::Vector2u w);
 	sf::Vector2i getpos();
 };
